requirements.txt
streamlit
google-cloud-vision
pillow
opencv-python-headless
pandas
openpyxl
